refactor: move earn calculations to helpers directory
