+++ features/earn/guni/open/containers/GuniOpenMultiplyVaultDetails.tsx
@@ -5,8 +5,7 @@ import { DetailsSection } from 'components/DetailsSection'
 import { DetailsSectionContentTable } from 'components/DetailsSectionContentTable'
 import { DetailsSectionFooterItemWrapper } from 'components/DetailsSectionFooterItem'
 import { ContentFooterItemsEarn } from 'components/vault/detailsSection/ContentFooterItemsEarn'
-import { calculateBreakeven } from 'features/earn/common/calculateBreakeven'
-import { calculateEarnings } from 'features/earn/common/calculateEarnings'
+import { calculateBreakeven, calculateEarnings } from 'helpers/earn/calculations'
 import { OAZO_LOWER_FEE } from 'helpers/multiply/calculations'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
